BUGZILLA-820102 Fix createSnapshot REST method (Manual cherry-pick of cd6d2c6e5c5e60753fcff09e9d753b937f43c4d7)

+++ guvnor-webapp/src/main/java/org/drools/guvnor/server/RepositoryPackageOperations.java
@@ -488,8 +488,10 @@ public class RepositoryPackageOperations {
         log.info( "USER:" + getCurrentUserName() + " CREATING PACKAGE SNAPSHOT for package: [" + packageName + "] snapshot name: [" + snapshotName );
 
         if ( replaceExisting ) {
-            getRulesRepository().removePackageSnapshot( packageName,
-                    snapshotName );
+            if (getRulesRepository().containsSnapshot(packageName, snapshotName)) {
+                getRulesRepository().removePackageSnapshot( packageName,
+                        snapshotName );
+            }
         }
 
         getRulesRepository().createPackageSnapshot( packageName,
